requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-environ==0.11.2
psycopg2==2.9.10
drf-yasg==1.21.8
django-cors-headers==4.6.0
